Add Bible book list and generating function
Add a bible book list (unique, ignoring numbers in front of books) of the books
in esvbible.  Add nonimplemented function to generate a regex which will match
every 3 character or more fragment that includes the start of a book.

E.g. Isiah will make a regex part that matches:
"Isiah", "Isia" and "Isi".

The three character lower limit is because esvbible does not return matches
for 2 and 1 characters.